heroku-deployment-future/requirements.txt
requests==2.26.0
flask==2.0.3
numpy==1.21.1
line-bot-sdk==2.1.0
urllib3==1.26.6
dateTime==4.3
matplotlib==3.4.2
typing-extensions==3.10.0.2
gunicorn==20.1.0